split files
uhhh, well I guess it do looks like something that might work, maybe
