+++ app.py
@@ -1,0 +1,6 @@
+from app import create_app, socketio
+
+app = create_app(debug=True)
+
+if __name__ == "__main__":
+    socketio.run(app, port=8000, debug=True)
