+++ app/static/css/dice/dice.css
