+++ app/static/fonts/font-awesome/fa-brands-400.ttf
