+++ app/static/fonts/font-awesome/fa-v4compatibility.woff2
