+++ app/static/js/dice.js
