+++ app/static/js/math.js
