+++ app/static/js/nukeops.js
