+++ app/templates/dice/dice.html
