+++ app/templates/main/main.html
@@ -3,7 +3,7 @@
 {% block content %}
 
 <div class="button-grid">
-    <a href="{{ url_for('main_ss13') }}">
+    <a href="{{ url_for('main.main_ss13') }}">
         <button class="redirect-button" id="neon"><i>SS13</i></button></a>
     <a href="https://discord.gg/bxgeRPq">
         <button class="redirect-button" id="neon"><i>DISCORD</i></button></a>
