+++ app/templates/main/rules.html
@@ -7,7 +7,7 @@
 </head>
 
 
-<a href="{{ url_for('main_ss13') }}">
+<a href="{{ url_for('main.main_ss13') }}">
 	<button><i>BACK</i>
 	</button>
 </a>
