+++ requirements.txt
@@ -2,4 +2,5 @@ flask==3.0.0
 Flask-SQLAlchemy==3.1.1
 flask_socketio==5.3.6
 gevent==23.9.1
+gevent-websocket==0.10.1
 mysqlclient==2.2.0
